+++ middleware.js
@@ -15,4 +15,4 @@ export default withAuth(
   },
 );
 
-export const config = { matcher: ['/app/*'] };
+export const config = { matcher: ['/app/:path*'] };
